Mise à jour des entités et du joueur

+++ include/attributs_statistiques.h
@@ -10,7 +10,7 @@
 
 
 
-int pv;             /**<  */
-int attaque;        /**<  */
-int defense;        /**<  */
-int vitesse;        /**<  */
+float pv;             /**<  */
+float attaque;        /**<  */
+float defense;        /**<  */
+float vitesse;        /**<  */
